feat: :tada: Added basic infra files along with first auth service.
Created basic infra files along with auth service. In auth service we created routers for signIn, signOut, signUp and to get current user. We implemented errors handling. where is 1 main abstract class for creating error handling with specific implementation and response type that is going to be used along whole app. There is one middleware for handling errors. Also we vreated connection with mongo db within one cluster(MicroServices approach(one db to service)) and we added typescript support to this module. Finaly we created service class Password for hashing. This mechanism is applied before saving user in to database in signUp